Restructure into more structs and add ssh functionality
To simplify the formatting and assignment, the PrePrompt struct was
broken up into two further structs that encompass SSH and VCS info
respectively. They are now optional members of the PrePrompt struct to
make it simple to format each and allow for easier tweaking and changes
in the future.

An SSH check has now been completed so that users are aware they are on
a remote connection.