Fetch hostname from hostname file instead of environment variable
Due to $HOSTNAME not being a posix defined variable it is often not
automatically defined/exported in the standard shell 'sh' where it often
is in bash or zsh. Since rust with run the process in sh, we should just
fetch the hostname from the hostname file in /etc/hostname instead of
hoping it will be there depending on the shell we are spawned in.

Also does some comment clean up to better align with rust's 100
character line length.