Highlight user name in white if root in SSH connections
Adds a check to see if the user is currently root during an SSH
connection, and highlights the name in white if so. Also adds some tests
around get_ssh_info(). All but on of the tests are commented out due to
`cargo test` executing in parallel and the tests changing environment
variables as if they are in an isolated process causing them to be
flaky.